sector size is more than just 512 bytes and 16-bit int is not enough, grrr

+++ bdos/fsdrive.c
@@ -173,7 +173,7 @@ long    log(BPB *b, int drv)
         OFD *fo,*f;                         /*  M01.01.03   */
         DND *d;
         DMD *dm;
-        int rsiz,cs,n,fs,ncl,fcl;
+        unsigned long rsiz,cs,n,fs,ncl,fcl;
 
         rsiz = b->recsiz;
         cs = b->clsiz;
